[parameter] move implementation of SimpleDict to header
